Add logic to the brain-even game
It calls a greeting function;
Generates and shows random number to a user and waits for a response
if the number is odd correct answer is 'yes' and 'no' otherwise.
After three correct answers the game succesfully finishes with a Congratulation message,
in any other case it results with a Wrong answer response and requests user to play again.